backend/requirements.txt
Flask
transformers
Pillow
torch
torchvision
gunicorn